BioMarkovChains.BMC

BioMarkovChains.BioMarkovChain

BioMarkovChains.dnaseqprobability

BioMarkovChains.generatedna

BioMarkovChains.hasprematurestop

BioMarkovChains.initials

BioMarkovChains.iscoding

BioMarkovChains.perronfrobenius

BioMarkovChains.transition_count_matrix

BioMarkovChains.transition_probability_matrix

BioMarkovChains.transitions

StatsAPI.fit!

BioMarkovChains.BMC — Type

BMC

Alias for the type BioMarkovChain.

BioMarkovChains.BioMarkovChain — Type

struct BioMarkovChain{M<:AbstractMatrix, I<:AbstractVector, N<:Integer} <: AbstractBioMarkovChain

A BioMarkovChain represents a Markov chain used in biological sequence analysis. It contains a transition probability matrix (tpm) and an initial distribution of probabilities (inits) and also the order of the Markov chain.

Fields

tpm::M: The transition probability matrix.

inits::I: The initial distribution of probabilities.

n::N: The order of the Markov chain.

Constructors

BioMarkovChain(tpm::M, inits::I, n::N=1) where {M<:AbstractMatrix, I<:AbstractVector, N<:Integer}: Constructs a BioMarkovChain object with the provided transition probability matrix, initial distribution, and order.

BioMarkovChain(sequence::LongNucOrView{4}, n::Int64=1): Constructs a BioMarkovChain object based on the DNA sequence and transition order.

Example

sequence = LongDNA{4}("ACTACATCTA")

model = BioMarkovChain(sequence, 2)
BioMarkovChain:
  - Transition Probability Matrix -> Matrix{Float64}(4 × 4):
    0.444    0.111	0.0	  0.444
    0.444    0.444	0.0	  0.111
    0.0      0.0	0.0	  0.0
    0.111    0.444	0.0	  0.444
  - Initial Probabilities -> Vector{Float64}(4 × 1):
    0.333
    0.333
    0.0
    0.333
  - Markov Chain Order:2

BioMarkovChains.dnaseqprobability — Method

sequenceprobability(sequence::LongNucOrView{4}, tpm::Matrix{Float64}, initials=Vector{Float64})

Compute the probability of a given sequence using a transition probability matrix and the initial probabilities distributions.

\[P(X_1 = i_1, \ldots, X_T = i_T) = \pi_{i_1}^{T-1} \prod_{t=1}^{T-1} a_{i_t, i_{t+1}}\]

Arguments

sequence::LongNucOrView{4}: The input sequence of nucleotides.

tm::BioMarkovChain is the actual data structure composed of a tpm::Matrix{Float64} the transition probability matrix and initials=Vector{Float64} the initial state probabilities.

Returns

probability::Float64: The probability of the input sequence.

Example

mainseq = LongDNA{4}("CCTCCCGGACCCTGGGCTCGGGAC")
   
bmc = BioMarkovChain(mainseq)

BioMarkovChain with DNA Alphabet:
  - Transition Probability Matrix -> Matrix{Float64}(4 × 4):
   0.0     1.0     0.0     0.0
   0.0     0.5     0.2     0.3
   0.25    0.125   0.625   0.0
   0.0     0.6667  0.3333  0.0
  - Initial Probabilities -> Vector{Float64}(4 × 1):
   0.087
   0.4348
   0.3478
   0.1304
  - Markov Chain Order -> Int64:
   1

newseq = LongDNA{4}("CCTG")

    4nt DNA Sequence:
    CCTG

dnaseqprobability(newseq, bmc)
    
    0.0217

BioMarkovChains.generatedna — Method

generatedna(pf::Matrix{Float64}, steps::Int64; extended_alphabet::Bool = false)

Generate a new DNA sequence using the Perron-Frobenius matrix-based approach.

This function utilizes the Perron-Frobenius matrix pf to generate a DNA sequence over a specified number of steps. Each step involves transitioning from the current DNA state to a new state based on the probabilities defined in the matrix pf.

Arguments

pf::Matrix{Float64}: The Perron-Frobenius matrix, a square matrix of transition probabilities. Each row corresponds to the source state, and each column corresponds to the probabilities of transitioning to different states.

steps::Int64: The number of steps to generate the new DNA sequence.

extended_alphabet::Bool = false: (Optional) If true, the function will use an extended DNA alphabet with additional characters beyond the standard 'A', 'C', 'G', and 'T'. Default is false.

Returns

A LongDNA{4} object representing the newly generated DNA sequence.

Example

using GeneFinder, BioSequences

seq = randdnaseq(10^6) # create a random DNA sequence of 1Mb
orfdna = getorfdna(seq, min_len=90)[1] # find a random ORF

pf = perronfrobenius(orfdna)

    4×4 Matrix{Float64}:
    0.333333  0.296296  0.238095  0.292683
    0.245614  0.314815  0.238095  0.317073
    0.210526  0.222222  0.238095  0.195122
    0.210526  0.166667  0.285714  0.195122

newdna = generatedna(pf, 100)

BioMarkovChains.hasprematurestop — Method

hasprematurestop(sequence::LongNucOrView{4})::Bool

Determine whether the sequence of type LongSequence{DNAAlphabet{4}} contains a premature stop codon.

Returns a boolean indicating whether the sequence has more than one stop codon.

BioMarkovChains.initials — Method

initials(sequence::SeqOrView{A}) where A

Calculate the estimated initial probabilities for a Markov chain based on a given sequence.

This function takes a sequence of states and calculates the estimated initial probabilities of each state in the sequence for a Markov chain. The initial probabilities are estimated by counting the occurrences of each state at the beginning of the sequence and normalizing the counts to sum up to 1.

\[\begin{align} \pi{i} = P(X_{i} = i), i \in T \\ \sum_{i=1}^{N} \pi_{i} = 1 \end{align}\]

Now using the dinucleotides counts estimating the initials would follow:

\[\hat{\pi_{i}} = c_{i} \sum_{k} c_{k}\]

Arguments

sequence::SeqOrView{A}: The sequence of states representing the Markov chain.

Returns

An Vector{Flot64} of estimated initial probabilities for each state in the sequence.

BioMarkovChains.iscoding — Function

iscoding(
    sequence::LongSequence{DNAAlphabet{4}}, 
    codingmodel::BioMarkovChain, 
    noncodingmodel::BioMarkovChain,
    η::Float64 = 1e-5
    )

Check if a given DNA sequence is likely to be coding based on a log-odds ratio. The log-odds ratio is a statistical measure used to assess the likelihood of a sequence being coding or non-coding. It compares the probability of the sequence generated by a coding model to the probability of the sequence generated by a non-coding model. If the log-odds ratio exceeds a given threshold (η), the sequence is considered likely to be coding. It is formally described as a decision rule:

\[S(X) = \log \left( \frac{{P_C(X_1=i_1, \ldots, X_T=i_T)}}{{P_N(X_1=i_1, \ldots, X_T=i_T)}} \right) \begin{cases} > \eta & \Rightarrow \text{{coding}} \\ < \eta & \Rightarrow \text{{noncoding}} \end{cases}\]

Arguments

sequence::LongSequence{DNAAlphabet{4}}: The DNA sequence to be evaluated.

codingmodel::BioMarkovChain: The transition model for coding regions.

noncodingmodel::BioMarkovChain: The transition model for non-coding regions.

η::Float64 = 1e-5: The threshold value (eta) for the log-odds ratio (default: 1e-5).

Returns

true if the sequence is likely to be coding.

false if the sequence is likely to be non-coding.

Raises

ErrorException: if the length of the sequence is not divisible by 3.

ErrorException: if the sequence contains a premature stop codon.

Example

sequence = LondDNA("ATGGCATCTAG")
codingmodel = BioMarkovChain()
noncodingmodel = BioMarkovChain()
iscoding(sequence, codingmodel, noncodingmodel)  # Returns: true

BioMarkovChains.perronfrobenius — Method

perronfrobenius(sequence::LongNucOrView{4}, n::Int64=1)

Compute the Perron-Frobenius matrix, a column-wise version of the transition probability matrix (TPM), for a given nucleotide sequence.

The Perron-Frobenius matrix captures the asymptotic probabilities of transitioning between nucleotides in the sequence over a specified number of steps n. It provides insight into the long-term behavior of a Markov chain or a dynamical system associated with the sequence.

Arguments

sequence::LongNucOrView{4}: A nucleotide sequence represented as a LongNucOrView{4} object.

n::Int64=1: The number of steps to consider for the transition probability matrix. Default is 1.

Returns

A copy of the Perron-Frobenius matrix. Each column of this matrix corresponds to the probabilities of transitioning from the current nucleotide state to all possible nucleotide states after n steps.

Example

```julia sequence = LongSequence{DNAAlphabet{4}}("ACGTCGTCCACTACGACATCAGC") # Replace with an actual nucleotide sequence n = 2 pf = perronfrobenius(sequence, n)

BioMarkovChains.transition_count_matrix — Method

transition_count_matrix(sequence::LongSequence{DNAAlphabet{4}})

Compute the transition count matrix (TCM) of a given DNA sequence.

Arguments

sequence::LongSequence{DNAAlphabet{4}}: a LongSequence{DNAAlphabet{4}} object representing the DNA sequence.

Returns

A Matrix object representing the transition count matrix of the sequence.

Example

seq = LongDNA{4}("AGCTAGCTAGCT")

tcm = transition_count_matrix(seq)

4×4 Matrix{Int64}:
 0  0  3  0
 0  0  0  3
 0  3  0  0
 2  0  0  0

BioMarkovChains.transition_probability_matrix — Method

transition_probability_matrix(sequence::LongSequence{DNAAlphabet{4}}, n::Int64=1)

Compute the transition probability matrix (TPM) of a given DNA sequence. Formally it construct $\hat{\mathscr{M}}$ where:

\[\mathscr{m}_{ij} = P(X_t = j \mid X_{t-1} = i) = \frac{{P(X_{t-1} = i, X_t = j)}}{{P(X_{t-1} = i)}}\]

The transition matrices of DNA and Amino-Acids are arranged sorted and in row-wise matrices:

First the DNA matrix:

\[\mathscr{M}_{DNA} = \begin{bmatrix} _{AA} & _{AC} & _{AG} & _{AT} \\ _{CA} & _{CC} & _{CG} & _{CT} \\ _{GA} & _{GC} & _{GG} & _{GT} \\ _{TA} & _{TC} & _{TG} & _{TT} \\ \end{bmatrix}\]

And then, the Aminoacids:

\[\mathscr{M}_{AA} = \begin{bmatrix} _{AA} & _{AC} & _{AD} & \dots & _{AW} \\ _{CA} & _{CC} & _{CD} & \dots & _{CW} \\ _{DA} & _{DC} & _{DD} & \dots & _{DW} \\ \vdots & \vdots & \vdots & \ddots & \vdots \\ _{WA} & _{WC} & _{WD} & \dots & _{WW} \\ \end{bmatrix}\]

Arguments

sequence::LongNucOrView{4}: a LongNucOrView{4} object representing the DNA sequence.

n::Int64=1: The order of the Markov model. That is the $\hat{M}^{n}$

Keywords

extended_alphabet::Bool=false: If true will pass the extended alphabet of DNA to search

Returns

A Matrix object representing the transition probability matrix of the sequence.

Example

seq = dna"AGCTAGCTAGCT"

tpm = transition_probability_matrix(seq)

4×4 Matrix{Float64}:
 0.0  0.0  1.0  0.0
 0.0  0.0  0.0  1.0
 0.0  1.0  0.0  0.0
 1.0  0.0  0.0  0.0

BioMarkovChains.transitions — Method

transitions(sequence::LongSequence)

Compute the transition counts of each pair in a given biological sequence sequence.

Arguments

sequence::LongSequence{DNAAlphabet{4}}: a LongSequence{DNAAlphabet{4}} object representing the DNA sequence.

Returns

A dictionary with keys being Dict{Tuple{DNA, DNA}, Int64} objects representing the dinucleotides, and values being the number of occurrences of each dinucleotide in the sequence.

Example

seq = dna"AGCTAGCTAGCT"

dinucleotides(seq)
Dict{Tuple{DNA, DNA}, Int64} with 4 entries:
  (DNA_C, DNA_T) => 3
  (DNA_G, DNA_C) => 3
  (DNA_T, DNA_A) => 2
  (DNA_A, DNA_G) => 3

StatsAPI.fit! — Method

fit!(bmc::BMC, inits:Vector{Float64}, tpm::Matrix{Float64})

Update bmc in-place based on information generated from a state sequence.